requirements.txt
# 🚀 IntegridAI HackAI 2025 - Python Dependencies
# Safe dependencies for hackathon development

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# AI Integration
openai==1.12.0
python-dotenv==1.0.0

# Data Processing  
pandas==2.0.3
numpy==1.24.3

# Development Tools
requests==2.31.0
asyncio  # Built-in with Python

# Database (SQLite for development)
sqlite3  # Built-in with Python

# Utilities
uuid  # Built-in with Python  
datetime  # Built-in with Python
json  # Built-in with Python
logging  # Built-in with Python
typing  # Built-in with Python
pathlib  # Built-in with Python
dataclasses  # Built-in with Python

# Testing (Optional)
pytest==7.4.2
pytest-flask==1.2.0

# Linting & Formatting (Optional)
black==23.7.0
flake8==6.0.0

# Documentation (Optional) 
sphinx==7.1.2